Modular code with Chatgpt.
I am improving the code with chatgpt in order to be able to analyse the data better. At this particular moment I am in a situation where I am not sure that backtesting_utils is actually adding any monrey to the lowest performing ETFs. I need to resume from them in the next session.