Remaking previous codes for the visualization of ETFs
Remaking of the codes Check_ETF_data.py and check_ETF_data_and_visualization. They are some simple codes that have some strategy analysis for ETFs. The strategy system is old and I am making a new and better one, however the plotting is ready and quite good. One of them also checks some small statistics of the ETFs, showing for example crosscorrelations with the S&P500.

Maybe I could make them some simple codes for analysing and plotting ETFs data. In the future some of the functionalities, especially the plotting will be integrated in the main modular program.